Create seperate books and rentals creation files
